Deactivate thread bar based on selections (#4755)
This change makes the thread bars gray when another thread is selected
or when it is outside of the time range.  I also changed
TicksToMicroseconds to make it possible to be negative if the time given
was before the start of the capture.

Screenshot: https://screenshot.googleplex.com/9i3ZXgcDQo9rUjT

Bug: http://b/240111360

+++ src/OrbitGl/include/OrbitGl/GlUtils.h
@@ -13,7 +13,7 @@
 }
 
 [[nodiscard]] inline double TicksToMicroseconds(uint64_t start, uint64_t end) {
-  return static_cast<double>(end - start) * 0.001;
+  return (static_cast<double>(end) - static_cast<double>(start)) * 0.001;
 }
 
 #endif  // ORBIT_GL_GL_UTILS_H_
